requirements.txt
streamlit>=1.39.0
mysql-connector-python>=8.2.0
pandas>=2.0.0
plotly>=5.18.0
bcrypt>=4.1.0
python-dotenv>=1.0.0
annotated-text>=0.3.0
openpyxl>=3.1.0
